rebase CMake and AAssetManager

+++ include/SaltyEngine/Maths.hpp
@@ -25,8 +25,8 @@ namespace SaltyEngine
         static float Sign(float value);
 
     public:
-        constexpr static const float rad2deg = (180 / M_PI);
-        constexpr static const float deg2rad = (M_PI / 180);
+        constexpr static const float rad2deg = static_cast<float>(180 / M_PI);
+        constexpr static const float deg2rad = static_cast<float>(M_PI / 180);
 
     private:
         static std::vector<float> m_sin;
